Deacoplated Git calls to a git helpers folder to be used by the AI
