+++ Synview-backend/utils/GITHelpers.ts
@@ -1,0 +1,54 @@
+import { Octokit } from "npm:@octokit/rest";
+const env = Deno.env.toObject();
+
+const GitHubToken = env.GITHUB_TOKEN;
+const octokit = new Octokit({
+  auth: GitHubToken,
+});
+
+export default async function diffExtracter(
+  owner: string,
+  repo: string,
+  ref: string
+): Promise<string> {
+  try {
+    const { data: diffs } = await octokit.request(
+      `GET /repos/{owner}/{repo}/commits/{ref}`,
+      {
+        owner: owner,
+        repo: repo,
+        ref: ref,
+        headers: {
+          accept: "application/vnd.github.diff",
+        },
+      }
+    );
+    return String(diffs);
+  } catch {
+    return "";
+  }
+}
+
+export async function GetMetadata(
+  owner: string,
+  repo: string,
+  sha: string
+): Promise<string> {
+  try {
+    const result = await octokit.request(
+      `GET /repos/{owner}/{repo}/commits/{ref}`,
+      {
+        owner: owner,
+        repo: repo,
+        ref: sha,
+      }
+    );
+    return JSON.stringify(result.data.commit);
+  } catch {
+    return "";
+  }
+}
+
+export function FileSearch(tool: string) {
+  return "not implemented yet ";
+}
